rename client_param to api_param
